+++ R/sample_data_funcs.R
@@ -71,6 +71,7 @@ rescale_var <- function(U, X, pars, family=1, link) {
     else if (family == 0 || family == 5) {
       if (link == "probit") Y <- 1*(eta + qnorm(U) > 0)
       else if (link == "logit") Y <- 1*(eta + qlogis(U) > 0)
+      else if (link == "log") Y <- 1*(eta + log(U) > 0)
       else stop("invalid link function for binomial distribution")
 
       # trunc <- pars$trunc
@@ -122,7 +123,7 @@ rescale_var <- function(U, X, pars, family=1, link) {
 ##' @describeIn rescale_var Old name, now deprecated
 ##' @export
 rescaleVar <- function(U, X, pars, family=1, link) {
-  deprecate_soft("0.8.0", "rescaleVar", "rescale_var")
+  deprecate_soft("0.8.0", "rescaleVar()", "rescale_var()")
   rescale_var(U, X, pars, family=family, link)
 }
 
@@ -197,7 +198,7 @@ rescale_cop <- function(U, X, beta, family=1, par2) {
   else if (family == 2) {
     # Y <- sqrt(phi)*qt(U, df=pars$par2) + eta
     param <- 2*expit(eta) - 1
-    Y <- cVCopula(U, copula = tCopula, par2=par2, param = param, inverse=TRUE)
+    Y <- cVCopula(U, copula = tCopula, param = param, par2=par2, inverse=TRUE)
   }
   else if (family == 3) {
     param <- exp(eta) - 1
@@ -226,7 +227,7 @@ rescale_cop <- function(U, X, beta, family=1, par2) {
 ##' @describeIn rescale_cop Old name, now deprecated
 ##' @export
 rescaleCop <- function(U, X, beta, family=1, par2) {
-  deprecate_soft("0.8.0", "rescaleCop", "rescale_cop")
+  deprecate_soft("0.8.0", "rescaleCop()", "rescale_cop()")
   rescale_cop(U, X, beta, family=family, par2)
 }
 
@@ -356,7 +357,7 @@ sim_copula <- function(dat, family, par, par2, model_matrix) {
 ##' @describeIn sim_copula Old name, now deprecated
 ##' @export
 sim_CopVal <- function(dat, family, par, par2, model_matrix) {
-  deprecate_soft("0.8.0", "sim_CopVal", "sim_copula")
+  deprecate_soft("0.8.0", "sim_CopVal()", "sim_copula()")
   sim_copula(dat, family, par, par2, model_matrix)
 }
 
@@ -382,16 +383,23 @@ glm_sim <- function (family, eta, phi, other_pars, link, quantiles=TRUE) {
   if (is.matrix(eta)) n <- nrow(eta)
   else n <- length(eta)
   if (missing(link)) {
-    ## get the default link for this family
-    link <- family_vals[family_vals$val==family,2]
+    if (is(family, "causl_family")) link <- family$link
+    else {
+      ## get the default link for this family
+      link <- family_vals[family_vals$val==family,2]
+    }
   }
 
   if (is(family, "causl_family")) {
     if (!(family$name %in% names(links_list))) {
-      if (!(link %in% family$links_list)) stop(paste0(link, " is not a valid link function for ", family$name, " family"))
-      stop(paste0("Family ", family$name, " is not a valid and registered family"))
+      ## this is not a registered family
+      # if (!(link %in% family$links_list)) stop(paste0(link, " is not a valid link function for ", family$name, " family"))
+      # stop(paste0("Family ", family$name, " is not a valid and registered family"))
+
+      if (!(link %in% unlist(links_list)) &&
+          !(link %in% names(family$custom_links))) stop("Link function not known")
     }
-    if (!(link %in% links_list[[family$name]])) stop(paste0(link, " is not a valid link function for ", family$name, " family"))
+    else if (!(link %in% links_list[[family$name]])) stop(paste0(link, " is not a valid link function for ", family$name, " family"))
 
     if (family$name %in% c("categorical","ordinal")) {
       if (ncol(eta) != other_pars$nlevel - 1) stop("Invalid 'eta' input to glm_sim")
@@ -406,12 +414,18 @@ glm_sim <- function (family, eta, phi, other_pars, link, quantiles=TRUE) {
       else stop("We shouldn't get here")
     }
     else {
-      if (link=="identity") mu <- eta
-      else if (link=="inverse") mu <- 1/eta
-      else if (link=="log") mu <- exp(eta)
-      else if (link=="logit") mu <- expit(eta)
-      else if (link=="probit") mu <- pnorm(eta)
-      else stop("We shouldn't get here")
+      if (link %in% unlist(links_list)) {
+        if (link=="identity") mu <- eta
+        else if (link=="inverse") mu <- 1/eta
+        else if (link=="log") mu <- exp(eta)
+        else if (link=="logit") mu <- expit(eta)
+        else if (link=="probit") mu <- pnorm(eta)
+        else stop("We shouldn't get here")
+      }
+      else {
+        if (link %in% names(family$custom_links)) mu <- family$custom_links[[link]]$linkinv(eta)
+        else stop("Link function not understood")
+      }
     }
 
     pars <- list(mu=mu)
@@ -473,6 +487,7 @@ glm_sim <- function (family, eta, phi, other_pars, link, quantiles=TRUE) {
     else if (family == 5) {
       if (link=="logit") mu <- expit(eta)
       else if (link=="probit") mu <- pnorm(eta)
+      else if (link=="log") mu <- exp(eta)
       else stop("Not a valid link function for the Bernoulli distribution")
 
       x <- rbinom(n, size=1, prob=mu)